+++ packages/@lou.codes/configs/eslint.config.js
@@ -1,7 +1,7 @@
 import { coreRules } from "./rules/core.js";
 import { functionalRules } from "./rules/functional.js";
 import { importRules } from "./rules/import.js";
-import { preferArrowRules } from "./rules/prefer-arrow.js";
+import { preferArrowFunctionsRules } from "./rules/prefer-arrow-functions.js";
 import { prettierRules } from "./rules/prettier.js";
 import { reactRules } from "./rules/react.js";
 import { typescriptRules } from "./rules/typescript.js";
@@ -15,7 +15,7 @@ export default /** @type {const} @satisfies {ReadonlyArray<import("eslint").Lint
 	coreRules,
 	...functionalRules,
 	importRules,
-	preferArrowRules,
+	preferArrowFunctionsRules,
 	prettierRules,
 	reactRules,
 	typescriptRules,