+++ packages/@lou.codes/configs/rules/prefer-arrow-functions.js
@@ -1,0 +1,28 @@
+import eslintPluginPreferArrowFunctions from "eslint-plugin-prefer-arrow-functions";
+import { freeze } from "../freeze.js";
+import { ERROR } from "./levels.js";
+
+/**
+ * ESLint `prefer-arrow` rules.
+ * @see [eslint-plugin-prefer-arrow-functions](https://npm.im/eslint-plugin-prefer-arrow-functions)
+ */
+export const preferArrowFunctionsRules = freeze(
+	/** @type {const} @satisfies {import("eslint").Linter.FlatConfig} */ ({
+		plugins: { "prefer-arrow-functions": eslintPluginPreferArrowFunctions },
+		rules: {
+			/**
+			 * Prefer arrow functions over traditional functions.
+			 */
+			"prefer-arrow-functions/prefer-arrow-functions": [
+				ERROR,
+				{
+					allowNamedFunctions: false,
+					classPropertiesAllowed: false,
+					disallowPrototype: false,
+					returnStyle: "implicit",
+					singleReturnOnly: false,
+				},
+			],
+		},
+	}),
+);